cost

+++ src/actions/IrEste.java
@@ -31,7 +31,7 @@ public class IrEste extends SearchAction{
 
 	@Override
 	public Double getCost() {
-		return new Double(1);
+		return new Double(10);
 	}
 
 	@Override
